QMYSQL c'est de la merde

+++ server/server/group.cpp
@@ -14,7 +14,7 @@ void group::setColor(const QColor &value){color = value;}
 
 void    group::addfils(int id)
 {
-    qDebug() << this->name.c_str() << " addfils: " << QString::number(id) << QString::number((int)this);
+    qDebug() << this->name.c_str() << " addfils: " << QString::number(id) << QString::number((long int)this);
     int dafuq = id;
     if (this->listfils.empty())
         qDebug() << "emptyfils";
